runtime/wasm: isolate scheduler state from native Gs
