+++ runtime/internal/runtime/fatal_default.go
@@ -1,0 +1,7 @@
+//go:build !llgo || !js || !wasm
+
+package runtime
+
+func fatal(s string) {
+	print("fatal error: ", s, "\n")
+}
